+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+
+- add `duration` parameter to `flyTo()`
+- fix `jumpTo()` never returns
+
 ## 0.0.1+1
 
 - fix urls to website